Refactor: update import paths for Kafka constants and event interfaces in cart module
